CRUD estudiantes 90% listo, bug crear mismo rut

+++ app/Http/Controllers/Admin/StudentController.php
@@ -17,7 +17,6 @@ use Freshwork\ChileanBundle\Rut;
 class StudentController extends Controller
 {
 
-
     /* security verification */
     public function __construct(){
         $this->middleware('auth');
@@ -34,6 +33,11 @@ class StudentController extends Controller
         return view('admin.students.index', compact('students'));
     }
 
+    public function index2($quantity_careers,$student)
+    {   
+        $data['quantity_careers'] = $quantity_careers;
+        return view('admin.students.create2',$data,compact('student'));
+    }
 
     /**
      * Show the form for creating a new resource.
@@ -46,46 +50,62 @@ class StudentController extends Controller
         return view('admin.students.create');
     }
 
+    public function create2(Request $request,$student){
+        return "hola bb";
+    }
     /**
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StudentStoreRequest $request)
+    public function store(Request $request)
     {
+
+        if($request->name == null or $request->email == null or $request->rut == null){
+            return redirect()->route('students.create')->with('info','¡ No deje datos en blanco  !');
+        }
+       
+        $career1 = $request->career1; //cada checkbox y lo comparo con cada carrera, si el checkbox no esta sleccionado sera vacio
+        $career2 = $request->career2;
+        $career3 = $request->career3;
+        $career4 = $request->career4;
+
+        if($career1 =="" and $career2 =="" and $career3 =="" and $career4 ==""){
+            return redirect()->route('students.create')->with('info','¡ Debe seleccionar al menos 1 carrera !');
+        }
+      
         $student = Student::create( $request->all() );
         $rut = $student->rut;
 
         $student->rut = strval(Rut::parse($rut)->number()).strval(Rut::parse($rut)->vn());
         $student->save();
        
-        
-        
+                    
         if (Rut::parse($rut)->quiet()->validate() == false ){//si la validacion es incorrecta borramos el estudiante
             $this->destroy($student->id);
             return redirect()->route('students.create')->with('info','¡ Rut mal ingresado, intente nuevamente !');
         }
         
-        $career = $student->career;
-        echo $career;
-        if($career == "Ingeniería en Computación e Informática"){
+       
+
+        if($career1 == "Ingeniería en Computación e Informática"){
             $student->careers()->attach(1);
         }
-        if($career == "Ingeniería Ejecución en Computación e Informática"){
+        if($career2 == "Ingeniería Ejecución en Computación e Informática"){
             $student->careers()->attach(2);
         }
 
-        if ($career == "Ingeniería Civil en Computación e Informática (Malla-Nueva)"){
+        if ($career3 == "Ingeniería Civil en Computación e Informática (Malla-Nueva)"){
             $student->careers()->attach(3);
         }
 
-        if ($career == "Ingeniería Civil en Computación e Informática (Malla-Antigua)"){
+        if ($career4 == "Ingeniería Civil en Computación e Informática (Malla-Antigua)"){
             $student->careers()->attach(4);
         }
         
-
-        return redirect()->route('students.index',$student->id)->with('info','Estudiante creado con éxito');
+        
+        return redirect()->route('students.index')->with('info','¡ Alumno Creado Con exito !');
     }
 
     /**
@@ -116,13 +136,15 @@ class StudentController extends Controller
      */
     public function edit($id)
     {
-        /*$works = StudentorderBy('id','ASC')->get();
-        $work=Studentfind($id);
-        $students = Student::orderBy('id','ASC')->get();
-        $academics = Academic::orderBy('id','ASC')->get();
-        $types = Type::orderBy('id','ASC')->get();
-        return view('admin.works.edit',compact('work','types','students','academics','works'));
-        */
+        $student =Student::find($id);
+        $rutconGuion = $student->rut;
+        $rutSinGuon=
+        $number = strVal(Rut::parse($rutconGuion)->number());
+        $vn = strVal(Rut::parse($rutconGuion)->vn());
+        $rutSinGuion=$number.$vn;
+
+        $rutFormateado = Rut::set()->number($number)->vn($vn)->format();
+        return view('admin.students.edit',compact('student','id','rutFormateado','rutSinGuion'));
     }
 
     /**
@@ -132,14 +154,44 @@ class StudentController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(WorkUpdateRequest $request, $id)
+    public function update(StudentUpdateRequest $request, $id)
     {
-        /*$work=Studentfind($id);
-        $work->fill($request->all())->save();
-        return  redirect()->route('works.index',$work->id)->with('info','Actividad de titulación actualizada correctamente');
-        */
-    }
+        $student=Student::find($id);
+        $rutStudent = $student->rut;
+        $request->rut=$rutStudent;
+        $student->careers()->detach(1);
+        $student->careers()->detach(2);
+        $student->careers()->detach(3);
+        $student->careers()->detach(4);//quitamos las relaciones viejas y para poner las nuevas
+        $student->fill($request->all())->save();
 
+        
+        $career1 = $request->career1; //cada checkbox y lo comparo con cada carrera, si el checkbox no esta sleccionado sera vacio
+        $career2 = $request->career2;
+        $career3 = $request->career3;
+        $career4 = $request->career4;
+
+        if($career1 =="" and $career2 =="" and $career3 =="" and $career4 ==""){
+            return redirect()->route('students.edit',$id)->with('info','¡ Debe seleccionar al menos 1 carrera !');
+        }
+
+        if($career1 == "Ingeniería en Computación e Informática"){
+            $student->careers()->attach(1);
+        }
+        if($career2 == "Ingeniería Ejecución en Computación e Informática"){
+            $student->careers()->attach(2);
+        }
+
+        if ($career3 == "Ingeniería Civil en Computación e Informática (Malla-Nueva)"){
+            $student->careers()->attach(3);
+        }
+
+        if ($career4 == "Ingeniería Civil en Computación e Informática (Malla-Antigua)"){
+            $student->careers()->attach(4);
+        }
+        return  redirect()->route('students.index',$student->id)->with('info',' Datos del Rut '.$rutStudent.' actualizado');
+        
+    }
     /**
      * Remove the specified resource from storage.
      *
@@ -151,8 +203,4 @@ class StudentController extends Controller
        Student::find($id)->delete();
        return back()->with('info','Eliminado correctamente');
     }
-
-
-    
-
 }
